Make the login form a component
